ISSUE #2345 - move addProjects
